+++ server/routes/users.js
@@ -102,13 +102,23 @@ export default (app) => {
                 return reply;
             }
 
+            const taskCreator = await app.objection.models.task.query().findOne({ creatorid: +req.params.id });
+            const taskExecutor = await app.objection.models.task.query().findOne({ executorid: +req.params.id });
+
+            if (taskCreator || taskExecutor) {
+                req.flash('info', i18next.t('flash.users.delete.task'));
+                reply.redirect(app.reverse('users'));
+
+                return reply;
+            }
+
             try {
                 await app.objection.models.userInfo.query().delete().where('user', +req.params.id);
                 await app.objection.models.user.query().deleteById(+req.params.id);
                 req.flash('info', i18next.t('flash.users.delete.success'));
                 reply.redirect(app.reverse('users'));
             } catch({ data }) {
-                req.flash('info', i18next.t('flash.users.delete.error'));
+                req.flash('error', i18next.t('flash.users.delete.error'));
                 reply.redirect(app.reverse('users'));
             }
